+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - upgrade to ember(-(cli|data))@~3.3.0
 - DX:
     - No more mirage fixtures
+    - Have guid-like IDs for mirage factories (nodes and users to start)
 
 ## [0.7.0] - 2018-08-07
 ### Added